batman-adv: Don't schedule OGM for disabled interface
A transmission scheduling for an interface which is currently dropped by
batadv_iv_ogm_iface_disable could still be in progress. The B.A.T.M.A.N. V
is simply cancelling the workqueue item in an synchronous way but this is
not possible with B.A.T.M.A.N. IV because the OGM submissions are
intertwined.

Instead it has to stop submitting the OGM when it detect that the buffer
pointer is set to NULL.

Reported-by: syzbot+a98f2016f40b9cd3818a@syzkaller.appspotmail.com
Reported-by: syzbot+ac36b6a33c28a491e929@syzkaller.appspotmail.com
Signed-off-by: Sven Eckelmann <sven@narfation.org>
---
Cc: Hillf Danton <hdanton@sina.com>

I had this one in the pipeline but didn't find the time to correctly test
it. You already send your one to syzkaller - do you want to rework your one
(to also catch the broken submission on the same interface) and submit it?
Or is this patch also ok for you? If it is, do you want to be mentioned in
some way or form in this patch?